# Optimize Hash#stringify_keys