--- a/activesupport/lib/active_support/json/encoding.rb
+++ b/activesupport/lib/active_support/json/encoding.rb
@@ -76,7 +76,13 @@ module ActiveSupport
             when Hash
               result = {}
               value.each do |k, v|
-                k = k.to_s unless String === k
+                unless String === k
+                  k = if Symbol === k
+                    k.name
+                  else
+                    k.to_s
+                  end
+                end
                 result[k] = jsonify(v)
               end
               result